apps/api/pyproject.toml
[project]
name = "codemeant-api"
version = "0.1.0"
description = "CodeMeant API - 프로젝트 가치 분석 서비스"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "supabase>=2.11.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.28.1",
    "openai>=1.58.1",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.3",
    "pydantic>=2.10.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
